+++ R/phenology_models.R
@@ -29,9 +29,7 @@ AT <- function(par, data){
   c <- par[5]
 
   # chilling
-  Rc <- data$Ti - T_base
-  Rc[Rc < 0] <- 1
-  Rc[Rc >= 0] <- 0
+  Rc <- ifelse(data$Ti >= T_base, 0, 1)
   Rc[1:t0,] <- 0
   Sc <- apply(Rc, 2, cumsum)